refactor: Complete Sequence 4 - Extract loop.py utilities to focused modules
Reduces loop.py from 958 → 640 lines (33% reduction) by extracting utilities
to separate, single-responsibility modules following SOLID principles.

**Files Created:**
- warmup.py - Warmup schedule utilities (get_focal_gamma)
- sampling.py - Balanced sampling for imbalanced datasets
- losses.py - FocalLoss implementation
- optimizer_factory.py - Optimizer and scheduler creation
- early_stopping.py - EarlyStopping handler

**Files Modified:**
- loop.py - Removed extracted code, updated imports (958 → 640 lines)
- __init__.py - Updated to expose utilities from new modules

**Benefits:**
- Single Responsibility: Each module has one focused purpose
- Easier Testing: Utilities can be tested in isolation
- Better Organization: Clear separation between orchestration and utilities
- Reduced Cognitive Load: Smaller files easier to understand

**Verification:**
- Unit tests: 10/10 pass ✅
- Integration tests: 19/19 pass ✅
- Type checks: mypy clean ✅
- Lint checks: ruff clean ✅

Completes Sequence 4 from EXECUTION_PLAN_2025-10-02.md
Part of REFACTORING_PLAN_V4.md Category 2
