+++ src/brain_brr/train/early_stopping.py
@@ -1,0 +1,45 @@
+"""Early stopping handler for training.
+
+Monitors validation metrics and stops training when no improvement is seen.
+"""
+
+from __future__ import annotations
+
+from src.brain_brr.config.schemas import EarlyStoppingConfig
+
+
+class EarlyStopping:
+    """Early stopping handler.
+
+    Encapsulates early stopping logic.
+    """
+
+    def __init__(self, config: EarlyStoppingConfig) -> None:
+        self.patience = config.patience
+        self.metric = config.metric
+        self.mode = config.mode
+        self.best_score = float("-inf") if self.mode == "max" else float("inf")
+        self.counter = 0
+        self.best_epoch = 0
+
+    def __call__(self, score: float, epoch: int = 0) -> bool:
+        """Check if should stop.
+
+        Args:
+            score: Current metric value
+            epoch: Current epoch
+
+        Returns:
+            True if should stop
+        """
+        improved = score > self.best_score if self.mode == "max" else score < self.best_score
+
+        if improved:
+            self.best_score = score
+            self.best_epoch = epoch
+            self.counter = 0
+            return False
+
+        self.counter += 1
+        # Allow exactly `patience` non-improving epochs; stop on the next one.
+        return self.counter > self.patience
