+++ src/brain_brr/train/loop.py
@@ -11,18 +11,13 @@ SOLID principles applied:
 from __future__ import annotations
 
 import logging
-import math
 import os
 from pathlib import Path
-from typing import Any, cast
+from typing import Any
 
 import torch
 import torch.multiprocessing as mp
-import torch.nn as nn
-import torch.nn.functional as tnf
-from torch.optim import AdamW, Optimizer
-from torch.optim.lr_scheduler import LambdaLR, LRScheduler
-from torch.utils.data import DataLoader, WeightedRandomSampler
+from torch.utils.data import DataLoader
 
 # Make TensorBoard optional
 try:
@@ -33,15 +28,12 @@ except ImportError:
     HAS_TENSORBOARD = False
     SummaryWriter = None  # type: ignore[assignment, misc]
 
-from src.brain_brr.config.schemas import (
-    Config,
-    EarlyStoppingConfig,
-    SchedulerConfig,
-    TrainingConfig,
-    WarmupScheduleConfig,
-)
+from src.brain_brr.config.schemas import Config
 from src.brain_brr.models import SeizureDetector
 from src.brain_brr.train.checkpoint import load_checkpoint, save_checkpoint
+from src.brain_brr.train.early_stopping import EarlyStopping
+from src.brain_brr.train.optimizer_factory import create_optimizer, create_scheduler
+from src.brain_brr.train.sampling import create_balanced_sampler
 from src.brain_brr.train.train_step import train_epoch
 from src.brain_brr.train.train_utils import set_seed, worker_init_fn
 from src.brain_brr.train.val_step import validate_epoch
@@ -57,317 +49,6 @@ LOG_EVERY_N_STEPS = int(os.getenv("BGB_LOG_EVERY_N_STEPS", "50"))
 # WSL2-safe multiprocessing defaults (must be before any DataLoader creation)
 if mp.get_start_method(allow_none=True) != "spawn":
     mp.set_start_method("spawn", force=True)
-
-# ============================================================================
-# Warmup Schedule Utilities (Production ML Best Practice)
-# ============================================================================
-
-
-def get_focal_gamma(
-    global_step: int,
-    warmup_config: WarmupScheduleConfig | None,
-    target_gamma: float = 2.0,
-) -> float:
-    """Compute focal loss gamma for current step.
-
-    Linear interpolation from start_gamma → end_gamma over warmup_steps.
-    Reduces loss amplification during early training for gradient stabilization.
-
-    Standard practice in production ML (OpenAI, Google, Meta).
-
-    Args:
-        global_step: Current training step
-        warmup_config: Warmup schedule configuration
-        target_gamma: Target gamma (from config focal_gamma)
-
-    Returns:
-        Effective gamma for current step
-    """
-    if warmup_config is None or not warmup_config.enabled or not warmup_config.focal_gamma_enabled:
-        return target_gamma
-
-    if global_step >= warmup_config.warmup_steps:
-        return target_gamma
-
-    # Linear interpolation: start → end over warmup_steps
-    progress = global_step / warmup_config.warmup_steps
-    start_gamma = warmup_config.focal_gamma_start
-    end_gamma = warmup_config.focal_gamma_end
-    current_gamma = start_gamma + progress * (end_gamma - start_gamma)
-
-    return current_gamma
-
-
-# ============================================================================
-# Reproducibility utilities (Single Responsibility)
-# ============================================================================
-
-
-# ============================================================================
-# Sampling utilities (Interface Segregation)
-# ============================================================================
-
-
-def create_balanced_sampler(dataset: Any, sample_size: int = 500) -> WeightedRandomSampler | None:
-    """Create positive-aware balanced sampler for imbalanced datasets.
-
-    Args:
-        dataset: EEGWindowDataset instance
-        sample_size: Number of windows to sample for statistics
-
-    Returns:
-        WeightedRandomSampler for balanced mini-batches or None if no seizures found
-    """
-    logger.info("[SAMPLER] Creating positive-aware balanced sampler...")
-
-    # Skip expensive sampling in smoke test mode
-    if env.smoke_test():
-        logger.info(
-            "[SMOKE TEST MODE] Skipping sampler window checking - returning None for uniform sampling"
-        )
-        return None
-
-    # Sample dataset to find which windows have seizures
-    sample_size = min(sample_size, len(dataset))
-    sample_indices = torch.randperm(len(dataset))[:sample_size]
-
-    # Track which windows actually have seizures
-    window_has_seizure = torch.zeros(len(dataset), dtype=torch.float32)
-    sampled_seizure_count = 0
-
-    logger.info(f"[SAMPLER] Checking {sample_size} windows for seizures...")
-    for i, idx in enumerate(sample_indices):
-        batch = dataset[idx.item()]
-        label = batch["label"]
-        if (label > 0).any():
-            window_has_seizure[idx] = 1.0
-            sampled_seizure_count += 1
-
-        # Progress update every 1000 windows
-        if (i + 1) % 1000 == 0:
-            logger.debug(
-                f"[SAMPLER] Checked {i + 1}/{sample_size} windows, found {sampled_seizure_count} with seizures"
-            )
-
-    # Estimate seizure ratio
-    seizure_ratio = sampled_seizure_count / sample_size
-    logger.info(
-        f"[SAMPLER] Final: {sampled_seizure_count}/{sample_size} windows with seizures ({seizure_ratio:.2%})"
-    )
-
-    if seizure_ratio < 1e-8:
-        logger.info("[SAMPLER] WARNING: No seizures found in sample! Using uniform sampling.")
-        return None
-
-    # Calculate weight for positive samples (sqrt to prevent explosion)
-    pos_weight = math.sqrt((1 - seizure_ratio) / seizure_ratio)
-
-    # Extrapolate to full dataset
-    # For unsampled indices, assign weight probabilistically
-    weights = torch.ones(len(dataset), dtype=torch.float32)
-
-    # Known seizure windows get high weight
-    weights[window_has_seizure > 0] = pos_weight
-
-    # Estimate weights for unsampled windows
-    unsampled_mask = torch.ones(len(dataset), dtype=torch.bool)
-    unsampled_mask[sample_indices] = False
-    n_unsampled_seizures = int(unsampled_mask.sum() * seizure_ratio)
-
-    if n_unsampled_seizures > 0:
-        unsampled_indices = torch.where(unsampled_mask)[0]
-        random_seizure_indices = unsampled_indices[
-            torch.randperm(len(unsampled_indices))[:n_unsampled_seizures]
-        ]
-        weights[random_seizure_indices] = pos_weight
-
-    logger.info(f"[SAMPLER] Seizure ratio: {seizure_ratio:.2%}")
-    logger.info(f"[SAMPLER] Positive weight: {pos_weight:.2f}")
-    logger.info(f"[SAMPLER] Estimated seizure windows: {(weights > 1).sum().item()}/{len(dataset)}")
-
-    return WeightedRandomSampler(
-        weights=weights.tolist(),
-        num_samples=len(weights),
-        replacement=True,
-        generator=torch.Generator().manual_seed(42),
-    )
-
-
-# ============================================================================
-# Loss functions (Open/Closed Principle)
-# ============================================================================
-
-
-class FocalLoss(nn.Module):
-    """Binary focal loss on logits with optional pos_weight.
-
-    This wraps BCE-with-logits and applies focal modulation:
-        loss = alpha_t * (1 - p_t)^gamma * BCEWithLogitsLoss(logits, targets)
-
-    - logits: (B, T)
-    - targets: (B, T) in {0,1}
-    - pos_weight: optional scalar tensor to up-weight positives (same semantics as BCEWithLogitsLoss)
-    """
-
-    def __init__(self, alpha: float = 0.25, gamma: float = 2.0) -> None:
-        super().__init__()
-        self.alpha = float(alpha)
-        self.gamma = float(gamma)
-
-    def forward(
-        self,
-        logits: torch.Tensor,
-        targets: torch.Tensor,
-        *,
-        pos_weight: torch.Tensor | None = None,
-    ) -> torch.Tensor:
-        # Per-element BCE on logits for numerical stability
-        # Clamp logits to prevent overflow in BCE computation
-        logits_clamped = logits.clamp(min=-100, max=100)
-        bce = tnf.binary_cross_entropy_with_logits(
-            logits_clamped, targets, reduction="none", pos_weight=pos_weight
-        )
-        # Probabilities (use clamped logits for numerical stability)
-        p = torch.sigmoid(logits_clamped)
-        # Critical: Clamp probabilities to avoid log(0) or log(1) issues
-        p = p.clamp(min=1e-6, max=1 - 1e-6)
-        p_t = p * targets + (1.0 - p) * (1.0 - targets)
-        # Class-balanced alpha
-        alpha_t = self.alpha * targets + (1.0 - self.alpha) * (1.0 - targets)
-        # Focal modulation with numerical stability
-        # Clamp p_t away from 1 to prevent (1-p_t)^gamma from underflowing to 0
-        p_t_stable = p_t.clamp(min=1e-7, max=1 - 1e-7)
-        mod = (1.0 - p_t_stable).pow(self.gamma)
-        focal_loss = alpha_t * mod * bce
-
-        # Additional safety: clamp output to prevent extreme values
-        focal_loss = focal_loss.clamp(max=100.0)  # Prevent loss explosion
-        return cast(torch.Tensor, focal_loss)
-
-
-# ============================================================================
-# Optimizer & Scheduler factories (Open/Closed Principle)
-# ============================================================================
-
-
-def create_optimizer(model: nn.Module, config: TrainingConfig) -> Optimizer:
-    """Create optimizer from config.
-
-    Factory pattern for optimizer creation.
-    Applies weight decay only to weights, not biases or normalization parameters.
-    """
-    if config.optimizer == "adamw":
-        # Separate parameters into decay and no_decay groups
-        # This prevents weight decay from corrupting normalization layers
-        no_decay = ["bias", "bn", "ln", "layernorm", "norm", "rmsnorm"]
-        decay_params = []
-        no_decay_params = []
-
-        for name, param in model.named_parameters():
-            if not param.requires_grad:
-                continue
-            # Check if parameter name contains any no_decay keyword
-            if any(nd in name.lower() for nd in no_decay):
-                no_decay_params.append(param)
-            else:
-                decay_params.append(param)
-
-        # Create parameter groups with different weight decay
-        param_groups = [
-            {
-                "params": decay_params,
-                "weight_decay": config.weight_decay,
-                "lr": config.learning_rate,
-            },
-            {"params": no_decay_params, "weight_decay": 0.0, "lr": config.learning_rate},
-        ]
-
-        logger.info("[OPTIMIZER] Created parameter groups:")
-        logger.info(f"  - Decay group: {len(decay_params)} parameters")
-        logger.info(f"  - No-decay group: {len(no_decay_params)} parameters")
-
-        return AdamW(param_groups, lr=config.learning_rate, betas=(0.9, 0.999), eps=1e-8)
-    else:
-        raise ValueError(f"Unknown optimizer: {config.optimizer}")
-
-
-def create_scheduler(
-    optimizer: Optimizer,
-    config: SchedulerConfig,
-    total_steps: int,
-) -> LRScheduler:
-    """Create learning rate scheduler.
-
-    Uses a LambdaLR for linear warmup followed by cosine decay.
-    Designed to step once per optimization update.
-    """
-    warmup_steps = max(1, int(config.warmup_ratio * total_steps))
-
-    if config.type == "cosine":
-        import math
-
-        # Preserve initial learning rates so creating the scheduler does not
-        # mutate optimizer.param_groups (some schedulers may do this).
-        initial_lrs = [g["lr"] for g in optimizer.param_groups]
-
-        def lr_lambda(step: int) -> float:
-            # Linear warmup
-            if step < warmup_steps:
-                return float(step + 1) / float(warmup_steps)
-            # Cosine decay to 0
-            progress = (step - warmup_steps) / max(1, (total_steps - warmup_steps))
-            return 0.5 * (1.0 + math.cos(math.pi * progress))
-
-        sched = LambdaLR(optimizer, lr_lambda=lr_lambda, last_epoch=-1)
-        # Reset any change at construction time.
-        for g, lr in zip(optimizer.param_groups, initial_lrs, strict=False):
-            g["lr"] = lr
-        return sched
-    else:
-        raise ValueError(f"Unknown scheduler: {config.type}")
-
-
-# ============================================================================
-# Early stopping (Single Responsibility)
-# ============================================================================
-
-
-class EarlyStopping:
-    """Early stopping handler.
-
-    Encapsulates early stopping logic.
-    """
-
-    def __init__(self, config: EarlyStoppingConfig) -> None:
-        self.patience = config.patience
-        self.metric = config.metric
-        self.mode = config.mode
-        self.best_score = float("-inf") if self.mode == "max" else float("inf")
-        self.counter = 0
-        self.best_epoch = 0
-
-    def __call__(self, score: float, epoch: int = 0) -> bool:
-        """Check if should stop.
-
-        Args:
-            score: Current metric value
-            epoch: Current epoch
-
-        Returns:
-            True if should stop
-        """
-        improved = score > self.best_score if self.mode == "max" else score < self.best_score
-
-        if improved:
-            self.best_score = score
-            self.best_epoch = epoch
-            self.counter = 0
-            return False
-
-        self.counter += 1
-        # Allow exactly `patience` non-improving epochs; stop on the next one.
-        return self.counter > self.patience
-
 
 # ============================================================================
 # Main training orchestrator (Dependency Inversion)
